#37 has_secure_password自作の修正
